Fix missing import; rename Connection folder
