+++ src/homematicip/auth.py
@@ -4,8 +4,8 @@ import logging
 import uuid
 from dataclasses import dataclass
 
-from homematicip.connection_v2.connection_context import ConnectionContext
-from homematicip.connection_v2.rest_connection import RestResult, RestConnection
+from homematicip.connection.connection_context import ConnectionContext
+from homematicip.connection.rest_connection import RestResult, RestConnection
 
 LOGGER = logging.getLogger(__name__)
