+++ src/homematicip/connection/buckets.py
