+++ src/homematicip/connection/client_characteristics_builder.py
