+++ src/homematicip/connection/connection_url_resolver.py
